requirements.txt
aiogram==3.13.1
groq==0.11.0
python-dotenv==1.0.1